+++ id_expl0rer/app/src/app.cpp
@@ -1,0 +1,11 @@
+//
+// Created by nohbu on 11/17/2022.
+//
+
+#include <engine/engine.h>
+#include "engine/app.h"
+
+app::app::app()
+{
+    mainLoop = new engine::mainLoop;
+}
