--- id_expl0rer/app/src/main.cpp
@@ -1,11 +1,0 @@
-
-
-#include <engine/engine.h>
-
-
-int main(int argc, char **argv)
-{
-    id_expl0rerEngine *game = new id_expl0rerEngine();
-    game->PrintHelloWorld();
-    return 0;
-}
