+++ id_expl0rer/engine/include/engine/app.h
@@ -1,0 +1,19 @@
+//
+// Created by nohbu on 11/17/2022.
+//
+
+#ifndef ID_EXPL0RER_DEVELOPMENT_APP_H
+#define ID_EXPL0RER_DEVELOPMENT_APP_H
+
+#include "engine.h"
+namespace app{
+    class app{
+        public:
+            app();
+            engine::mainLoop* mainLoop;
+
+    };
+}
+
+
+#endif //ID_EXPL0RER_DEVELOPMENT_APP_H
