+++ id_expl0rer/engine/include/engine/engine.h
@@ -1,6 +1,16 @@
 #pragma once
+#include <time.h>
 
-class engine {
-    public:
-        static void PrintHelloWorld();
-};
+namespace engine {
+    class mainLoop {
+        public:
+            mainLoop();
+            void update();
+            bool running;
+            long long frame;
+            long long milliseconds;
+            long long deltatime;
+        private:
+            long long start;
+    };
+}
