+++ id_expl0rer/engine/include/engine/player.h
@@ -1,0 +1,18 @@
+//
+// Created by nohbu on 11/17/2022.
+//
+
+#ifndef ID_EXPL0RER_DEVELOPMENT_PLAYER_H
+#define ID_EXPL0RER_DEVELOPMENT_PLAYER_H
+
+namespace player {
+    class renderer{
+    public:
+        int DisplayWidth;
+        int DisplayHeight;
+    };
+
+}
+
+#endif //ID_EXPL0RER_DEVELOPMENT_PLAYER_H
+
